Move service deliveries into its own sub-app
Continuing with the sub-apps for features work, this moves service
deliveries and its concerns into one folder.
