+++ src/apps/service-deliveries/services/index.js
@@ -1,0 +1,7 @@
+const data = require('./services/data.service')
+const formatting = require('./services/formatting.service')
+
+module.exports = {
+  data,
+  formatting,
+}
