+++ src/services/company.service.js
@@ -4,7 +4,7 @@ const logger = require('../../config/logger')
 const adviserRepository = require('../repos/adviser.repo')
 const companyRepository = require('../repos/company.repo')
 const metadataRepository = require('../repos/metadata.repo')
-const serviceDeliveryRepository = require('../repos/service-delivery.repo')
+const serviceDeliveryRepository = require('../apps/service-deliveries/service-deliveries.repo')
 const interactionDataService = require('./interaction-data.service')
 const { getFormattedAddress } = require('../lib/address')
