+++ test/unit/services/service-delivery.service.test.js
@@ -1,4 +1,4 @@
-const servicedeliveryservice = require('~/src/services/service-delivery.service')
+const servicedeliveryservice = require('~/src/apps/service-deliveries/services/data.service')
 
 describe('Service delivery formatting service', function () {
   describe('convertServiceDeliveryFormToApiFormat', function () {
